% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breast.R
\docType{data}
\name{breast}
\alias{breast}
\title{Breast Cancer}
\format{
A data frame with 286 rows and 10 variables:
\describe{
  \item{\code{age}}{factor. 20-29, 30-39, 40-49, 50-59, 60-69, 70-79.}
  \item{\code{menopause}}{factor.  lt40, ge40, premeno.}
  \item{\code{tumor.size}}{factor.  0-4, 5-9, 10-14, 15-19, 20-24, 25-29, 30-34, 35-39, 40-44, 45-49, 50-54.}
  \item{\code{inv.nodes}}{factor. 0-2, 3-5, 6-8, 9-11, 12-14, 15-17, 18+.}
  \item{\code{node.caps}}{factor. yes, no.}
  \item{\code{deg.malig}}{factor. 1, 2, 3. Higher numbers indicate greater malignancy.}
  \item{\code{breast}}{factor.  left, right.}
  \item{\code{breast.quad}}{factor.  left-up, left-low, right-up, right-low, central.}
  \item{\code{irradiate}}{factor. yes, no.}
  \item{\code{recurrence}}{factor. yes, no.}
}
}
\source{
This breast cancer domain was obtained from the University Medical 
Centre, Institute of Oncology, Ljubljana, Yugoslavia. 
Thanks go to M. Zwitter and M. Soklic for providing the data. 
Downloaded from \href{https://www.openml.org/d/13}{OpenML}.
}
\usage{
breast
}
\description{
Breast cancer data
}
\note{
\code{recurrence} is the response or outcome variable.
}
\examples{
summary(breast)
}
\keyword{datasets}
