% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contraception.R
\docType{data}
\name{contraception}
\alias{contraception}
\title{Contraceptive method choice}
\format{
A data frame with 1473 rows and 10 variables:
\describe{
  \item{\code{wife_age}}{integer.}
  \item{\code{wife_education}}{integer. \code{1=low, 2, 3, 4=high}.}
  \item{\code{number_of_children_ever_born}}{integer. }
  \item{\code{wife_religion}}{factor. \code{Non-Islam, Islam}.}
  \item{\code{standard_of_living}}{integer. \code{1=low, 2, 3, 4=high}.}
  \item{\code{media_exposure}}{factor. \code{Good, Not Good}.}
  \item{\code{contraceptive}}{factor. \code{No-use, Long-term, Short-term}.}
  \item{\code{wife_now_working}}{factor. \code{Yes, No}.}
  \item{\code{husband_education}}{integer. \code{1=low, 2, 3, 4=high.}}
  \item{\code{husband_occupation}}{integer. \code{1,2,3,4}.}
}
}
\source{
Dua, D. and Graff, C. (2019).
UCI Machine Learning Repository [http://archive.ics.uci.edu/ml].
Irvine, CA: University of California, School of Information and Computer Science
}
\usage{
contraception
}
\description{
This dataset is a subset of the 1987 National Indonesia
Contraceptive Prevalence Survey.
}
\details{
The samples are married women
who were either not pregnant or do not know if they were at the
time of interview. The problem is to predict the current
contraceptive method choice (no use, long-term methods,
or short-term methods) of a woman based on her demographic and
socio-economic characteristics.
}
\examples{
summary(contraception)
}
\keyword{datasets}
