% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corplot.R
\name{corplot}
\alias{corplot}
\title{Correlation Matrix}
\usage{
corplot(data, sort = FALSE, lab_size = 4)
}
\arguments{
\item{data}{data frame}

\item{sort}{logical. If \code{TRUE}, reorder variables to place variables
with similar correlation patterns together.}

\item{lab_size}{size for correlation coefficient labels (default=4).}
}
\value{
a ggplot graph
}
\description{
Create a correlation matrix for all quantitative variables in a data frame.
}
\details{
The \code{corplot} function will only select quantitative variables from
a data frame. Categorical variables are ignored.

The correlation matrix is presented as a lower triangle matrix.
Correlations are Pearson Product Moment correlation coefficients.

Missing values are deleted in listwise fashion.
}
\note{
This function is a wrapper for the \code{\link[ggcorrplot:ggcorrplot]{ggcorrplot}} function.
}
\examples{
corplot(cars74)
corplot(cars74, sort=TRUE)
corplot(coffee, lab_size=2, sort=TRUE)
}
