% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Two-way frequency table}
\usage{
crosstab(
  data,
  rowvar,
  colvar,
  type = c("freq", "cellpercent", "rowpercent", "colpercent"),
  total = TRUE,
  na.rm = TRUE,
  digits = 2,
  plot = FALSE
)
}
\arguments{
\item{data}{data frame}

\item{rowvar}{row factor (unquoted)}

\item{colvar}{column factor (unquoted)}

\item{type}{statistics to print. Options are \code{"freq"}, \code{"cellpercent"},
\code{"rowpercent"}, or \code{"colpercent"} for frequencies, cell percents,
row percents, or column percents).
Default: \code{"freq"}.}

\item{total}{if TRUE, includes total percents, Default: TRUE}

\item{na.rm}{if TRUE, deletes cases with missing values, Default: TRUE}

\item{digits}{number of decimal digits to report for percents, Default: 2}

\item{plot}{logical. If \code{TRUE} generate stacked bar chart. Default: FALSE.}
}
\value{
a list with 9 components:
\itemize{
\item{\code{tbl}}{ (table). Table of frequencies}
\item{\code{cellPcts}}{ (table). Table of cell percents}
\item{\code{rowPcts}}{ (table). Table of row percents}
\item{\code{colPcts}}{ (table). Table of column percents}
\item{\code{type}}{ (character). Type of table to print}
\item{\code{total}}{ (logical). If \code{TRUE}, print row and or column totals}
\item{\code{digits}}{ (numeric). number of digits to print}
\item{\code{rowname}}{ (character). Row variable name}
\item{\code{colname}}{ (character). Column variable name}
}
}
\description{
This function creates a two way frequency table.
}
\details{
This function calculates and prints a two way frequency table.
Given a data frame, a row factor, a column factor, and a type (frequencies, cell percents,
row percents, or column percents) the function prints a table with labeled rows
and columns, frequencies or percents with the percentage sign depending on
type, a level labeled <NA> if na.rm = FALSE, and a level labeled Total if
total = TRUE.
}
\examples{
crosstab(mtcars, cyl, gear)

crosstab(mtcars, cyl, gear, type="cellpercent")

crosstab(mtcars, cyl, gear, type="colpercent")

crosstab(mtcars, cyl, gear, type="rowpercent")
}
