% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_summary.R
\name{df_summary}
\alias{df_summary}
\title{df_summary}
\usage{
df_summary(df, digits = 2)
}
\arguments{
\item{df}{a data frame}

\item{digits}{– number of decimal digits for statistics, Default: 2}
}
\value{
the data data frame invisibly (so that it can be used in a pipeline)
}
\description{
A comprehensive description of a data frame
}
\details{
prints a comprehensive description of a data frame via several tables,
a general summary table and tables that provide a breakdown of quantitative and categorical variables
}
\examples{
data<-data.frame("height"=c(4,5,3,2, 100), 'weight'=c(39,88,NA,15, -2),"names"=c('Bill',"Dean", "Sam", NA, "Jane"), 'race'=c('b','w','w','o', 'b'))
df_summary(mtcars)
}
