% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfPlot.R
\name{dfPlot}
\alias{dfPlot}
\title{Visualize a data frame}
\usage{
dfPlot(data)
}
\arguments{
\item{data}{a data frame.}
}
\value{
a \code{ggplot2} graph
}
\description{
\code{dfPlot} visualizes the variables in a data frame.
}
\details{
For each variable, the plot displays
\itemize{
  \item{type (\code{numeric},
              \code{integer},
              \code{factor},
              \code{ordered factor},
              \code{logical}, or \code{date})}
  \item{percent of available (and missing) cases}
}
Variables are sorted by type and the total number of variables
and cases are printed in the caption.
}
\examples{
dfPlot(coffee)
dfPlot(pbp2018)
}
\seealso{
For more descriptive statistics on
a data frame see \link{df_summary}.
}
