% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa.R
\name{FA}
\alias{FA}
\title{Factor analysis}
\usage{
FA(data, nfactors = NULL, rotate = "none", fm = "pa", digits = 2, ...)
}
\arguments{
\item{data}{a data frame or correlation matrix.}

\item{nfactors}{nuber of factors to extract.}

\item{rotate}{factor rotation to perform.}

\item{fm}{type of factor extraction.}

\item{digits}{number of digits to retain.}

\item{...}{parameters passed to the \code{psych::fa} function.}
}
\value{
returns a list with 5 components:
\item{call}{the call}
\item{loadings}{factor pattern}
\item{variance}{variance accounted for}
\item{Structure}{structure matrix}
\item{phi}{factor intercorrelations for oblique rotations}
\item{scores}{factor scores if factors are extracted from a data frame}
}
\description{
\code{FA} performs common factor analysis
}
\details{
The \code{FA} function is a wrapper for the \code{psych:fa} function.
Factor extractions include principal axis (\code{pa}) and maximum likelihood
(\code{ml}), among others. Factor rotations include \code{none}, \code{varimax},
and \code{promax}.
}
\examples{
fit.fa <- FA(Harman74.cor$cov, nfactors=4, rotate="varimax")
plot(fit.fa)
plot(fit.fa, type="bar")
}
\seealso{
\link{PCA}, \link{plot.factorAnalysis}, \link{score}, and
\link{scree_plot}.
}
