% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupdiff.R
\name{groupdiff}
\alias{groupdiff}
\title{Test of group differences}
\usage{
groupdiff(
  formula,
  data,
  method = c("anova", "kw"),
  digits = 2,
  offset = 5,
  horizontal = FALSE
)
}
\arguments{
\item{formula}{formula of the form \code{Y ~ X} where \code{Y} is the response
variable, and \code{X} is the categorical explanatory variable.}

\item{data}{a data frame.}

\item{method}{character. Either \code{"anova"}, or \code{"kw"} (see details).}

\item{digits}{Number of significant digits to print.}

\item{offset}{space at top of graph for post-hoc comparison letters.}

\item{horizontal}{logical. It \code{TRUE}, boxplots are plotted horizontally.}
}
\value{
a list with 2 components
}
\description{
One-way analysis (ANOVA or Kruskal-Wallis Test)
with post-hoc comparisons and plots
}
\details{
The \code{groupdiff} function performs one of two analyses:
\describe{
\item{\code{anova}}{A one-way analysis of variance, with TukeyHSD
post-hoc comparisons.}
\item{\code{kw}}{A Kruskal Wallis Rank Sum Test, with Conover
Test post-hoc comparisons.}
}
In each case, a boxplot and summary statistics are provided.
}
\examples{
# parametric analysis
groupdiff(hp ~ gear, mtcars)

# nonparametric analysis
groupdiff(popularity ~ vehicle_style, cardata, method="kw",
          offset=500, horizontal=TRUE)
}
\seealso{
\link[PMCMRplus]{kwAllPairsConoverTest},
\link[multcompView]{multcompLetters}.
}
