% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holland.R
\docType{data}
\name{holland}
\alias{holland}
\title{Holland Occupational Themes}
\format{
A data frame with 18231 rows and 68 variables:
\describe{
 \item{\code{R1}}{Test the quality of parts before shipment}
 \item{\code{R2}}{Lay brick or tile}
 \item{\code{R3}}{Work on an offshore oil-drilling rig}
 \item{\code{R4}}{Assemble electronic parts}
 \item{\code{R5}}{Operate a grinding machine in a factory}
 \item{\code{R6}}{Fix a broken faucet}
 \item{\code{R7}}{Assemble products in a factory}
 \item{\code{R8}}{Install flooring in houses}
 \item{\code{I1}}{Study the structure of the human body}
 \item{\code{I2}}{Study animal behavior}
 \item{\code{I3}}{Do research on plants or animals}
 \item{\code{I4}}{Develop a new medical treatment or procedure}
 \item{\code{I5}}{Conduct biological research}
 \item{\code{I6}}{Study whales and other types of marine life}
 \item{\code{I7}}{Work in a biology lab}
 \item{\code{I8}}{Make a map of the bottom of an ocean}
 \item{\code{A1}}{Conduct a musical choir}
 \item{\code{A2}}{Direct a play}
 \item{\code{A3}}{Design artwork for magazines}
 \item{\code{A4}}{Write a song}
 \item{\code{A5}}{Write books or plays}
 \item{\code{A6}}{Play a musical instrument}
 \item{\code{A7}}{Perform stunts for a movie or television show}
 \item{\code{A8}}{Design sets for plays}
 \item{\code{S1}}{Give career guidance to people}
 \item{\code{S2}}{Do volunteer work at a non-profit organization}
 \item{\code{S3}}{Help people who have problems with drugs or alcohol}
 \item{\code{S4}}{Teach an individual an exercise routine}
 \item{\code{S5}}{Help people with family-related problems}
 \item{\code{S6}}{Supervise the activities of children at a camp}
 \item{\code{S7}}{Teach children how to read}
 \item{\code{S8}}{Help elderly people with their daily activities}
 \item{\code{E1}}{Sell restaurant franchises to individuals}
 \item{\code{E2}}{Sell merchandise at a department store}
 \item{\code{E3}}{Manage the operations of a hotel}
 \item{\code{E4}}{Operate a beauty salon or barber shop}
 \item{\code{E5}}{Manage a department within a large company}
 \item{\code{E6}}{Manage a clothing store}
 \item{\code{E7}}{Sell houses}
 \item{\code{E8}}{Run a toy store}
 \item{\code{C1}}{Generate the monthly payroll checks for an office}
 \item{\code{C2}}{Inventory supplies using a hand-held computer}
 \item{\code{C3}}{Use a computer program to generate customer bills}
 \item{\code{C4}}{Maintain employee records}
 \item{\code{C5}}{Compute and record statistical and other numerical data}
 \item{\code{C6}}{Operate a calculator}
 \item{\code{C7}}{Handle customers' bank transactions}
 \item{\code{C8}}{Keep shipping and receiving records}
  \item{\code{TIPI1}}{Extraverted, enthusiastic.}
  \item{\code{TIPI2}}{Critical, quarrelsome.}
  \item{\code{TIPI3}}{Dependable, self-disciplined.}
  \item{\code{TIPI4}}{Anxious, easily upset.}
  \item{\code{TIPI5}}{Open to new experiences, complex.}
  \item{\code{TIPI6}}{Reserved, quiet.}
  \item{\code{TIPI7}}{Sympathetic, warm.}
  \item{\code{TIPI8}}{Disorganized, careless.}
  \item{\code{TIPI9}}{Calm, emotionally stable.}
  \item{\code{TIPI10}}{Conventional, uncreative.}
  \item{\code{education}}{factor. How much education have you completed?}
  \item{\code{urban}}{factor. What type of area did you live when you were a child?}
  \item{\code{gender}}{factor. What is your gender?}
  \item{\code{age}}{integer. Age in years.}
  \item{\code{hand}}{factor. What hand do you use to write with?}
  \item{\code{religion}}{factor. What is your religion?}
  \item{\code{voted}}{factor. Have you voted in a national election in the past year?}
  \item{\code{married}}{factor. What is your marital status?.}
  \item{\code{familysize}}{integer. Including you, how many children did your mother have?}
  \item{\code{major}}{factor. If you attended a university, what was your major?}
}
}
\source{
Dataset obtained from the
\href{https://openpsychometrics.org/_rawdata/}{Open-Source Psychometrics Project}.
}
\usage{
holland
}
\description{
RIASEC/Holland Code test
}
\details{
This is a nonrandom subset of the larger online data collected
between 2015-2018.
The sample has been restricted to a US sample with the most commonly
identified majors and religions. Not all variables were selected.

The RIASEC has 48 items, rated on a 1-5 scale of how much they would
like to perform that task, with the labels 1=Dislike, 3=Neutral, 5=Enjoy.
The form 6 scales:
\describe{
\item{Realistic}{Doer - prefers technical, outdoor and athletic pursuits}
\item{Investigative}{Thinker - prefers scientific, research and intellectual pursuits}
\item{Artistic}{Creator - prefers creative, imaginative and intuitive pursuits}
\item{Social}{Helper - prefers helping, and developing and interpersonal pursuits}
\item{Enterprising}{Persuader - prefers leadership, influencing and persuasive pursuits}
\item{Conventional}{Organizer - prefers data management, numerical and organizational pursuits}
}

TIPI1 through TIPI10 are the Ten Item Personality Inventory
(see Gosling, S. D., Rentfrow, P. J., & Swann, W. B., Jr. (2003).
A Very Brief Measure of the Big Five Personality Domains.
Journal of Research in Personality, 37, 504-528.).

The TIPI items were rated "I see myself as:" _____ from
1 = Disagree strongly to 7=Agree strongly.
}
\examples{
summary(holland)
}
\keyword{datasets}
