% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift.R
\name{lift_plot}
\alias{lift_plot}
\title{Lift and gain charts}
\usage{
lift_plot(actual, prob, positive = "yes")
}
\arguments{
\item{actual}{actual class.}

\item{prob}{predicted probability of target class.}

\item{positive}{which class is the target class?}
}
\value{
a data frame with lift and gain values for each decile.
}
\description{
Create lift and gain charts for binary
classification problems
}
\examples{
# logistic regression example
data(heart)
heart <- na.omit(heart)
fit <- glm(disease ~ . , heart, family=binomial)
pred <- predict(fit, newdata=heart, type="response")
results <- lift_plot(heart$disease, pred, positive="yes")
print(results)
}
