% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{PCA}
\alias{PCA}
\title{Principal Components Analysis}
\usage{
PCA(data, nfactors = NULL, rotate = "none", digits = 2, ...)
}
\arguments{
\item{data}{a data frame or correlation matrix.}

\item{nfactors}{nuber of factors to extract.}

\item{rotate}{factor rotation to perform.}

\item{digits}{number of digits to retain.}

\item{...}{parameters passed to the \code{psych::principal} function.}
}
\value{
returns a list with 5 components:
\item{call}{the call}
\item{loadings}{structure matrix}
\item{variance}{variance accounted for}
\item{phi}{component intercorrelations for oblique rotations}
\item{scores}{component scores if factors are extracted from a data frame}
}
\description{
\code{PCA} performs a principal components analysis
}
\details{
The \code{PCA} function is a wrapper for the \code{psych::principal} function.
Component rotations include \code{none}, \code{varimax},
and \code{promax}.
}
\examples{
fit.pca <- PCA(Harman74.cor$cov, nfactors=4, rotate="varimax")
plot(fit.pca)
plot(fit.pca, type="bar")
}
\seealso{
\link{FA}, \link{plot.factorAnalysis}, \link{score}, and
\link{scree_plot}.
}
