% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.factorAnalysis.R
\name{plot.factorAnalysis}
\alias{plot.factorAnalysis}
\title{Plot a Factor Solution}
\usage{
\method{plot}{factorAnalysis}(x, sort = TRUE, type = c("table", "bar"), ...)
}
\arguments{
\item{x}{an object of class \code{factorAnalysis} produced by the \code{\link{PCA}}
or \code{\link{FA}} functions.}

\item{sort}{logical. If \code{TRUE}, sort the pattern matrix.}

\item{type}{generate a table plot (\code{"table"}) or bar plot (\code{"bar"}).}

\item{...}{not currently used}
}
\value{
a ggplot2 graph
}
\description{
This function plots the pattern matrix for a principal components or
common factor analysis solution
}
\examples{
fit.pca <- PCA(Harman74.cor$cov, nfactors=4, rotate="varimax")
plot(fit.pca)
plot(fit.pca, type="bar")
}
\seealso{
\link{PCA} and \link{FA}.
}
