% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.tab.R
\name{print.tab}
\alias{print.tab}
\title{Print a tab object}
\usage{
\method{print}{tab}(df, ...)
}
\arguments{
\item{...}{Parameters passed to a function}

\item{x}{An object of class \code{tab}}
}
\description{
Print the results of calculating a frequency table
}
\examples{
\dontrun{
frequency <- tab(venues, type, sort = TRUE, na.rm = FALSE)
print(frequency)
}
}
