% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot.R
\name{profile_plot}
\alias{profile_plot}
\title{Cluster profile plots}
\usage{
profile_plot(data, cluster = "cluster", type = c("bar", "line"))
}
\arguments{
\item{data}{a data frame.}

\item{cluster}{variable containing cluster membership.}

\item{type}{character. \code{"bar"} or \code{"line"} plots.}
}
\value{
a ggplot2 graph
}
\description{
\code{profile_plot} generates a mean profile plot for each cluster in
a cluster analysis.
}
\details{
The data frame must contain numeric variables for all variables except
the cluster variable. The cluster variable can be numeric, character, or
factor. The data will be standardized before plotting. The bar charts
are faceted, while the line charts are grouped.
}
\examples{
# 3 cluster solution
iris <- qacr::standardize(iris)
df <- iris[-5]
set.seed(1234)
fit <- kmeans(df, 3)
df$cluster <- fit$cluster
profile_plot(df, type="bar")
profile_plot(df, type="line")
}
