% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qacD-package.R, R/qacr-package.R
\docType{package}
\name{qacr-package}
\alias{qacr}
\alias{qacr-package}
\title{qacr: Functions to Facilitate Data Science Projects}
\description{
This package provides functions for descriptive statistics,
  data management, and data visualization.

This package provides functions for descriptive statistics,
  data management, and data visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rkabacoff/qacr}
  \item Report bugs at \url{https://github.com/rkabacoff/qacr/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/rkabacoff/qacr}
  \item Report bugs at \url{https://github.com/rkabacoff/qacr/issues}
}

}
\author{
\strong{Maintainer}: QAC 356 \email{rkabacoff@wesleyan.edu}


\strong{Maintainer}: QAC 356 \email{rkabacoff@wesleyan.edu}

}
\keyword{internal}
