% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qstats.r
\name{qstats}
\alias{qstats}
\title{Summary statistics for a quantitative variable}
\usage{
qstats(data, x, ...)
}
\arguments{
\item{data}{data frame}

\item{x}{numeric variable in data (unquoted)}

\item{...}{list of grouping variables}
}
\value{
a data frame, where columns are grouping variables (optional) and
statistics
}
\description{
This function provides descriptive statistics for a quantitative
variable alone or seperately by groups. You can provide user-defined statistics.
}
\details{
By default, \code{dstats} calculates the n, mean, and sd of the
numeric variable. However you can select any statistics that return
single numbers. For example, adding
\code{statistics=c("n", "median", "min", "max")} returns
those statistics instead.

Be default, cases with missing values on x or the grouping variables are
deleted. Add \code{na.rm = FALSE} to change this behavior.

By default, two significant digits are printed. Add
\code{digits=3} to print 3 significant digits, etc.
}
\examples{
# If no keyword arguments are provided, default values are used
qstats(mtcars, mpg, am, gear)

# You can supply as many (or no) grouping variables as needed
qstats(mtcars, mpg)

qstats(mtcars, mpg, am, gear, cyl, carb)

# You can input user-defined functions
# my_n <- function(xs) length(xs)
# qstats(mtcars, mpg, stats = c("my_n"), am, gear)
}
