% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{review}
\alias{review}
\title{review}
\usage{
review(df, digits = 2, maxcat = 10, label_length = 40)
}
\arguments{
\item{df}{a data frame}

\item{digits}{number of decimal digits for statistics.}

\item{maxcat}{maximum number of levels of a character/factor variable to print.}

\item{label_length}{maximum length of factor level label to print. Longer labels
will be truncated.}
}
\value{
the data frame invisibly (so that it can be used in a pipeline)
}
\description{
A comprehensive description of a data frame
}
\details{
Prints a comprehensive description of a data frame via several tables,
a general summary table and tables that provide a breakdown of quantitative and categorical variables
}
\examples{
testdata <- data.frame(height=c(4, 5, 3, 2, 100),
                       weight=c(39, 88, NA, 15, -2),
                       names=c("Bill","Dean", "Sam", NA, "Jane"),
                       race=c('b', 'w', 'w', 'o', 'b'))

review(testdata)

}
