% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocPlot.R
\name{rocPlot}
\alias{rocPlot}
\title{ROC Plot}
\usage{
rocPlot(model, positive, n.cuts = 20, labelsize = 3, labelround = 2)
}
\arguments{
\item{model}{predictive model from the caret train function.}

\item{positive}{label for the class to predict.}

\item{n.cuts}{number of probability cutpoints to plot.}

\item{labelsize}{size of cutpoint labels.}

\item{labelround}{number of decimal digits in the cutpoint labels.}
}
\value{
a ggplot2 graph
}
\description{
Plot a receiver operating curve for a binary predictive model
}
\details{
Takes a model produced by the train function in the caret package and
the binary level to predict, and generates a ROC plot.
}
\examples{

# logistic regression example
library(caret)
data(GermanCredit)
model.lr <- train(Class ~ Age + Amount + Duration,
                  data=GermanCredit,
                  method="glm", family="binomial")
rocPlot(model=model.lr, positive="Bad")

}
