% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{Scatterplot}
\usage{
scatter(data, x, y, fit = TRUE, jitter = FALSE, color = "deepskyblue2")
}
\arguments{
\item{data}{data frame}

\item{x}{quantitative predictor variable}

\item{y}{quantitative response variable}

\item{fit}{logical. If \code{TRUE}, add a line
of best fit with 95 percent confidence interval.}

\item{jitter}{logical. If \code{TRUE}, points
are jittered}

\item{color}{character. color for points}
}
\value{
a ggplot graph
}
\description{
Create a scatter plot between two
quantitative variables.
}
\details{
The \code{scatter} function creates a scatter plot with
a linear line of best fit. The equation and R-squared,
along with level of significance.
}
\examples{
scatter(cars74, wt, mpg)

}
