% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screePlot.R
\name{screePlot}
\alias{screePlot}
\title{Scree Test}
\usage{
screePlot(data, n.iter = 50, method = c("pc", "pa", "ml"), ...)
}
\arguments{
\item{data}{a data frame or correlation matrix.}

\item{n.iter}{number of iterations for parallel analysis.}

\item{method}{factoring method (\code{pc}, \code{pa}, or \code{ml}).}

\item{...}{parameters passed to \code{psych::fa.parallel}.}
}
\value{
a ggplot2 graph
}
\description{
\code{screePlot} generates a scree plot with superimpose parallel analysis
}
\details{
The \code{screePlot} function is a wrapper for the \code{psych::parallel} function.
A ggplot2 graph is produced with the scree plot and parallel analysis (scree plot for
simulated data) based on the method chosen. The methods are principal components ("pc"),
principal axis ("pa"), and maximum likelihood ("ml") factoring.
}
\examples{
screePlot(Harman74.cor$cov, method="pc", n.obs=145)
}
