% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{stats}
\alias{stats}
\title{stats}
\usage{
stats(data, x, statistics = c("n", "mean", "sd"), na.rm = TRUE,
  digits = 2, ...)
}
\arguments{
\item{data}{data frame}

\item{x}{numeric variable in data (unquoted)}

\item{statistics}{statistics to calculate (any function that produces a
value), Default: c("n", "mean", "sd")}

\item{na.rm}{if TRUE, delete cases with missing values on x and or grouping
variables, Default: TRUE}

\item{digits}{number of decimal digits to print, Default: 2}

\item{...}{list of grouping variables}
}
\value{
a data frame, where columns are grouping variables (optional) and statistics
}
\description{
Descriptive statistics for a quantitative variable
}
\examples{
stats(mtcars, mpg, c("n", "mean", "sd"), na.rm = TRUE, digits = 0, am, gear)
}
