% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Frequency distribution for a categorical variable}
\usage{
tab(
  data,
  x,
  sort = FALSE,
  maxcat = NULL,
  minp = NULL,
  na.rm = FALSE,
  total = FALSE,
  digits = 2,
  cum = FALSE,
  plot = FALSE
)
}
\arguments{
\item{data}{A dataframe}

\item{x}{A factor in the dataframe}

\item{sort}{Sort levels from high to low, Default: FALSE}

\item{maxcat}{Maximum number of categories to be included.
Smaller categories will be combined into an "Other" category.
Default: NULL}

\item{minp}{Minimum proportion for a category to be included.
Categories
representing smaller proportions willbe combined into an
"Other" category.
maxcat and minp cannot both be specified. Default: NULL}

\item{na.rm}{Removes missing values when TRUE, Default: FALSE}

\item{total}{Include a total category when true, Default: FALSE}

\item{digits}{Number of digits the percents should be rounded to, Default: 2}

\item{cum}{logical. If \code{TRUE}, include cumulative counts
and percents. In this case \code{total} will be set to \code{FALSE}.}

\item{plot}{logical. If \code{TRUE}, generate bar chart for frequency table.
Default: FALSE}
}
\value{
A data frame
}
\description{
Function to calculate frequency distributions
for categorical variables
}
\details{
The function \code{tab} will calculate the frequency
distribution
for a categorical variable and output a data.frame with three
columns:
level, n, percent.
}
\examples{
tab(venues, state, sort = TRUE, na.rm = TRUE,
    maxcat = 10, digits = 3)

tab(cars74, carb, cum=TRUE, plot=TRUE)
}
