% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wss_plot.R
\name{wss_plot}
\alias{wss_plot}
\title{WSS Plot}
\usage{
wss_plot(data, nc = 15, standardize = TRUE, seed = 1234)
}
\arguments{
\item{data}{a data frame of numeric variables.}

\item{nc}{integer. The largest number of clusters to evaluate.}

\item{standardize}{logical. If \code{TRUE}, standardize data before
clustering.}

\item{seed}{integer. A random number seed for reproducablility.}
}
\value{
a ggplot2 graph
}
\description{
Within Groups Sum of Squares Plot
}
\details{
\code{wss_plot} generates a plot of within-groups
sums-of-squares vs. number of clusters based on
k-means clustering. The clustering uses euclidean distances
between observations. By default, the variables are standardized
(recommended). The plot is useful for determining the number of clusters
present in the data. Look for the point were adding clusters does not significantly
decrease the within-groups sum-of-squares.
}
\examples{
# iris example (should find 3 clusters)
data(iris)
wss_plot(iris[-5])
}
