% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{crosstab}
\alias{crosstab}
\title{Two Way Frequency Table}
\usage{
crosstab(data, rowvar, colvar, type = "freq", total = TRUE,
  na.rm = FALSE, digits)
}
\arguments{
\item{data}{data frame}

\item{rowvar}{row factor (unquoted)}

\item{colvar}{column factor (unquoted)}

\item{type}{table contains one of:c("freq", "rowpercent", "colpercent")
representing frequencies, cell percents, row percents, or column percents,
Default: "freq"}

\item{total}{if TRUE, includes total percents, Default: TRUE}

\item{na.rm}{if TRUE, deletes cases with missing values, Default: FALSE}

\item{digits}{number of decimal digits to report for percents, Default: 2}
}
\value{
two way frequency table
}
\description{
This function creates a two way frequency table.
}
\details{
This function creates a two way frequency table that returns a table
printed with rows and columns, percents with the percentage sign, a level
labeled <NA> if na.rm = FALSE, and a level labeled <Total> if total = TRUE.
This function coerces x into a factor if it is not already a factor.
}
\examples{
crosstab(mtcars, cyl, gear, type="freq", digits=2)
crosstab(mtcars, cyl, gear, type="rowpercent", digits=3)
}
