% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qacr-package.R
\docType{package}
\name{qacr-package}
\alias{qacr}
\alias{qacr-package}
\title{qacr: Functions to Facilitate Data Science Projects}
\description{
This package provides functions for descriptive statistics,
  data management, and data visualization.
}
\author{
\strong{Maintainer}: Robert Kabacoff \email{rkabacoff@wesleyan.edu}

}
\keyword{internal}
