% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_plot.R
\name{roc_plot}
\alias{roc_plot}
\title{Receiver Operating Characteristic curve}
\usage{
roc_plot(actual, prob, positive, n.cuts = 20, labelsize = 3, labelround = 2)
}
\arguments{
\item{actual}{actual class.}

\item{prob}{predicted probability for the target class}

\item{positive}{label for the target class}

\item{n.cuts}{number of probability cut-points to plot}

\item{labelsize}{size of cutpoint labels}

\item{labelround}{number of decimal digits in the cutpoint labels}
}
\value{
a ggplot2 graph
}
\description{
Plot a receiver operating characteristic curve for a binary predictive model
}
\examples{
# logistic regression
data(heart)
heart <- na.omit(heart)
fit <- glm(disease ~ . , heart, family=binomial)
prob <- predict(fit, newdata=heart, type="response")
roc_plot(heart$disease, prob, positive="yes")
}
