% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.r
\name{stats}
\alias{stats}
\title{stats}
\usage{
stats(data, x, ...)
}
\arguments{
\item{data}{data frame}

\item{x}{numeric variable in data (unquoted)}

\item{...}{list of grouping variables}

\item{statistics}{statistics to calculate (any function that produces a
value), Default: c("n", "mean", "sd")}

\item{na.rm}{if TRUE, delete cases with missing values on x and or grouping
variables, Default: TRUE}

\item{digits}{number of decimal digits to print, Default: 2}
}
\value{
a data frame, where columns are grouping variables (optional) and
statistics
}
\description{
This function provides descriptive statistics for a quantitative
variable. You can provide user-defined statistics.
}
\examples{
# If no keyword arguments are provided, default values are used
stats(mtcars, mpg, am, gear)

# You can supply as many (or no) grouping variables as needed
stats(mtcars, mpg)
stats(mtcars, mpg, am, gear, cyl, carb)

# You can input user-defined functions
my_n <- function(xs) length(xs)
stats(mtcars, mpg, statistics = c("my_n"), am, gear)
}
