% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Frequency Distribution for a categorical variable}
\usage{
tab(data, x, sort = FALSE, maxcat = NULL, minp = NULL,
  na.rm = FALSE, total = FALSE, digits = 2)
}
\arguments{
\item{data}{A dataframe}

\item{x}{A factor in the dataframe}

\item{sort}{Sort levels from high to low, Default: FALSE}

\item{maxcat}{Maximum number of categories to be included.
Will combine extra categories into an other category.  Default: NULL}

\item{minp}{Minimum proportion for a category, combines extra categories into
an other category. Cannot specifiy both maxcat and minp. Default: NULL}

\item{na.rm}{Removes missing values when TRUE, Default: FALSE}

\item{total}{Include a total category when true, Default: FALSE}

\item{digits}{Number of digits the percents should be rounded to, Default: 2}
}
\value{
A data.frame
}
\description{
FUNCTION_DESCRIPTION
}
\details{
The function \code{tab} will calculate the frequency distribution
for a categorical variable and output a data.frame with three columns:
level, n, percent.
}
\examples{
\dontrun{
frequency <- tab(venues, state, sort = TRUE, na.rm = TRUE, maxcat = 6, digits = 3)
print(frequency)
}
}
